Update tools and doc for raspberry pi os (#2)
* Update tools and documentation for Raspberry Pi OS

This is the first of a series of changes to make the tools and
documentation work with the later Raspberry Pi OS.

We need to replace the deprecated raspistill tool with rpicam-apps, and
use Python 3 instead of Python 2.7.

* Fix typos etc

Some follow-up changes.

* Refactor code: python2 -> python3

Add requirements.txt as well.

* Update README.md

Need to apt install some python deps.

* Use BytesIO to replace StringIo

For Python3.

* xrange -> range

Python2 to Python3 transition.

* Some more refactoring

Refactor code and add --dry-run option.

* Fix dry-run logic

Fix bug.

* Better dry run conditions

Improve dry-run conditions.

* Print main loop iteration count

More diagnostic information.

* More prints

Add more stdout info.

* Use tweepy to replace twython

This is for API 1.1 support.

* Attempt to fix 403 error

Use a tweepy client for this.

* Update README.md

Mention the creation of an X project for the app.

* Bug fix

Fix a bug in tweepy client invocation.

* Improve twitimg-motion-run.sh

* Update README.md etc

And install python packages in virtual env.

* Bugfix: python venv

Fix a bug in script.